refactor: rename `NotificationRect` to `BannerRect` and `NotificationStack` to `WindowManager`
